Rework how utils are organized due to Polyglot overloading limitations
