+++ build.sbt
@@ -5,8 +5,11 @@ import sbt.Keys.{libraryDependencies, scalacOptions}
 import sbt.addCompilerPlugin
 import sbt.complete.DefaultParsers._
 import sbt.complete.Parser
-import sbtcrossproject.CrossPlugin.autoImport.{CrossType, crossProject}
-import src.main.scala.licenses.{DistributionDescription, SBTDistributionComponent}
+import sbtcrossproject.CrossPlugin.autoImport.{crossProject, CrossType}
+import src.main.scala.licenses.{
+  DistributionDescription,
+  SBTDistributionComponent
+}
 
 import java.io.File
 
@@ -14,9 +17,9 @@ import java.io.File
 // === Global Configuration ===================================================
 // ============================================================================
 
-val scalacVersion = "2.13.8"
-val graalVersion = "21.3.0"
-val javaVersion = "11"
+val scalacVersion         = "2.13.8"
+val graalVersion          = "21.3.0"
+val javaVersion           = "11"
 val defaultDevEnsoVersion = "0.0.0-dev"
 val ensoVersion = sys.env.getOrElse(
   "ENSO_VERSION",
@@ -713,11 +716,11 @@ lazy val `profiling-utils` = project
       "org.netbeans.api" % "org-netbeans-modules-sampler" % netbeansApiVersion
       exclude ("org.netbeans.api", "org-openide-loaders")
       exclude ("org.netbeans.api", "org-openide-nodes")
-      exclude("org.netbeans.api", "org-netbeans-api-progress-nb")
-      exclude("org.netbeans.api", "org-netbeans-api-progress")
-      exclude("org.netbeans.api", "org-openide-util-lookup")
-      exclude("org.netbeans.api", "org-openide-util")
-      exclude("org.netbeans.api", "org-openide-dialogs")
+      exclude ("org.netbeans.api", "org-netbeans-api-progress-nb")
+      exclude ("org.netbeans.api", "org-netbeans-api-progress")
+      exclude ("org.netbeans.api", "org-openide-util-lookup")
+      exclude ("org.netbeans.api", "org-openide-util")
+      exclude ("org.netbeans.api", "org-openide-dialogs")
       exclude ("org.netbeans.api", "org-openide-filesystems")
       exclude ("org.netbeans.api", "org-openide-util-ui")
       exclude ("org.netbeans.api", "org-openide-awt")
@@ -1006,7 +1009,6 @@ val truffleRunOptions = if (java.lang.Boolean.getBoolean("bench.compileOnly")) {
     "-Dpolyglot.engine.BackgroundCompilation=false"
   )
 }
-
 
 val truffleRunOptionsSettings = Seq(
   fork := true,
@@ -1744,7 +1746,8 @@ lazy val `std-base` = project
     Compile / packageBin / artifactPath :=
       `base-polyglot-root` / "std-base.jar",
     libraryDependencies ++= Seq(
-      "com.ibm.icu" % "icu4j" % icuVersion
+      "com.ibm.icu"         % "icu4j"       % icuVersion,
+      "org.graalvm.truffle" % "truffle-api" % graalVersion % "provided"
     ),
     Compile / packageBin := Def.task {
       val result = (Compile / packageBin).value
@@ -1767,7 +1770,7 @@ lazy val `std-table` = project
     Compile / packageBin / artifactPath :=
       `table-polyglot-root` / "std-table.jar",
     libraryDependencies ++= Seq(
-      "com.ibm.icu"         % "icu4j"             % icuVersion % "provided",
+      "com.ibm.icu"         % "icu4j"             % icuVersion   % "provided",
       "com.univocity"       % "univocity-parsers" % "2.9.1",
       "org.apache.poi"      % "poi-ooxml"         % "5.2.2",
       "org.apache.xmlbeans" % "xmlbeans"          % "5.1.0",
