+++ std-bits/base/src/main/java/org/enso/base/time/TimeUtilsBase.java
@@ -1,0 +1,3 @@
+package org.enso.base.time;
+
+public interface TimeUtilsBase {}
