+++ test/Tests/src/Data/Time/Date_Spec.enso
@@ -1,9 +1,8 @@
 from Standard.Base import all
 
 import Standard.Base.Data.Text.Text_Sub_Range
-import Standard.Base.Data.Time.Duration
 import Standard.Base.Data.Time.Date_Period
-import Standard.Base.Data.Time.Time_Period
+import Standard.Base.Data.Time.Duration
 from Standard.Base.Error.Common import Time_Error
 
 import Standard.Test
@@ -134,10 +133,7 @@ spec_with name create_new_date parse_date =
             date_1>date_2 . should_be_true
             date_1<date_2 . should_be_false
 
-    Date_Part_Spec.spec name create_new_date
-
-    Test.group name+" start_of/end_of period" <|
-        Test.specify "should allow to find start of a Date_Period containing the current date" <|
+        Test.specify "should allow to find start and end of a Date_Period containing the current date" <|
             d1 = create_new_date 2022 9 12
             d1.start_of Date_Period.Year . should_equal (Date.new 2022 1 1)
             d1.end_of Date_Period.Year . should_equal (Date.new 2022 12 31)
@@ -179,6 +175,9 @@ spec_with name create_new_date parse_date =
 
             (create_new_date 2000 7 1).end_of Date_Period.Quarter . should_equal (Date.new 2000 9 30)
             (create_new_date 2000 6 30).end_of Date_Period.Quarter . should_equal (Date.new 2000 6 30)
+
+    Date_Part_Spec.spec name create_new_date
+
 
 main = Test.Suite.run_main spec
